refactor(characters): extract Characters.Resolution
Seventh slice of #134. auto_resolve_pending and random_resolve_all share
the resolve_all loop (unified in #127), differing only in their
resolvable? predicate and value method, so they move together as the
random-resolution concern. Kept separate from Progression to hold both
under the ~400-line target: Progression computes what is pending,
Resolution decides it.
